+++ chat/message-feed/src/MessageFeed/MessageFeed.types.ts
@@ -3,5 +3,5 @@ import React from 'react';
 import { DarkModeProps } from '@leafygreen-ui/lib';
 
 export interface MessageFeedProps
-  extends React.ComponentProps<'div'>,
+  extends React.ComponentPropsWithRef<'div'>,
     DarkModeProps {}