+++ chat/message-rating/src/MessageRating/index.ts
@@ -1,6 +1,4 @@
 export { MessageRating } from './MessageRating';
-/** @deprecated */
-export * as lgMessageRatingStyles from './MessageRating.styles';
 export {
   type MessageRatingProps,
   MessageRatingValue,