+++ chat/message-rating/src/RadioButton/index.ts
@@ -1,4 +1,0 @@
-export { RadioButton } from './RadioButton';
-/** @deprecated */
-export * as lgRadioButtonStyles from './RadioButton.styles';
-export type { RadioButtonProps } from './RadioButton.types';