+++ chat/message-rating/src/index.ts
@@ -1,15 +1,5 @@
 export {
-  /** @deprecated */
-  lgMessageRatingStyles,
   MessageRating,
   type MessageRatingProps,
   MessageRatingValue,
 } from './MessageRating';
-export {
-  /** @deprecated */
-  lgRadioButtonStyles,
-  /** @deprecated */
-  RadioButton,
-  /** @deprecated */
-  type RadioButtonProps,
-} from './RadioButton';