+++ chat/title-bar/src/TitleBar/TitleBar.styles.ts
@@ -1,29 +1,25 @@
 import { css, cx } from '@leafygreen-ui/emotion';
 import { Theme } from '@leafygreen-ui/lib';
-import { palette } from '@leafygreen-ui/palette';
-import { spacing } from '@leafygreen-ui/tokens';
+import {
+  color,
+  InteractionState,
+  spacing,
+  Variant,
+} from '@leafygreen-ui/tokens';
 
-import { Align } from './TitleBar.types';
-
-const baseStyles = css`
+const getBaseTitleBarStyles = (theme: Theme) => css`
   width: 100%;
+  border-bottom: 1px solid
+    ${color[theme].border[Variant.Secondary][InteractionState.Default]};
+  background-color: ${color[theme].background[Variant.Primary][
+    InteractionState.Default
+  ]};
+  padding: 14px ${spacing[400]}px;
   display: flex;
-  justify-content: space-between;
+  justify-content: center;
   align-items: center;
-  padding: 14px ${spacing[4]}px;
-  border-bottom: 1px solid;
+  gap: ${spacing[200]}px;
 `;
-
-const themeStyles: Record<Theme, string> = {
-  [Theme.Dark]: css`
-    background-color: ${palette.black};
-    border-color: ${palette.gray.dark2};
-  `,
-  [Theme.Light]: css`
-    background-color: ${palette.white};
-    border-color: ${palette.gray.light2};
-  `,
-};
 
 export const getTitleBarStyles = ({
   className,
@@ -32,21 +28,5 @@ export const getTitleBarStyles = ({
   theme: Theme;
   className?: string;
 }) => {
-  return cx(baseStyles, themeStyles[theme], className);
+  return cx(getBaseTitleBarStyles(theme), className);
 };
-
-const contentContainerStyles = css`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: ${spacing[200]}px;
-`;
-
-const contentAlignmentStyles = css`
-  margin: auto;
-`;
-
-export const getContentContainerStyles = ({ align }: { align: Align }) =>
-  cx(contentContainerStyles, {
-    [contentAlignmentStyles]: align === Align.Center,
-  });