+++ chat/title-bar/src/TitleBar/index.ts
@@ -1,0 +1,2 @@
+export { TitleBar } from './TitleBar';
+export { type TitleBarProps } from './TitleBar.types';