+++ chat/message-feedback/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @lg-chat/message-feedback
 
+## 9.0.0
+
+### Major Changes
+
+- 3dfa899: [LG-5575](https://jira.mongodb.org/browse/LG-5575)
+
+  - Added compatibility with `@lg-chat/leafygreen-chat-provider@6.0.0`. We recommend new projects use `@lg-chat/leafygreen-chat-provider` v6; support for v5 will be removed in a future major version.
+  - Removed `PopoverMessageFeedback` component and `PopoverMessageFeedbackProps` type exports.
+
+### Patch Changes
+
+- 9cf3b18: Updates provider peer dependency version string to correctly use `pnpm` `workspace` syntax
+- Updated dependencies [9cf3b18]
+- Updated dependencies [3dfa899]
+  - @leafygreen-ui/icon-button@17.1.3
+  - @leafygreen-ui/typography@22.2.2
+  - @leafygreen-ui/text-area@12.1.3
+  - @leafygreen-ui/button@25.1.3
+  - @lg-chat/leafygreen-chat-provider@6.0.0
+
 ## 8.1.0
 
 ### Minor Changes