+++ chat/message-prompts/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @lg-chat/message-prompts
 
+## 4.1.4
+
+### Patch Changes
+
+- 9cf3b18: Updates provider peer dependency version string to correctly use `pnpm` `workspace` syntax
+- Updated dependencies [9cf3b18]
+  - @leafygreen-ui/icon-button@17.1.3
+  - @leafygreen-ui/typography@22.2.2
+
+## 4.1.3
+
+### Patch Changes
+
+- 68b59a0: [LG-5687](https://jira.mongodb.org/browse/LG-5687): update border color from `green.dark1` to `green.dark2`
+
 ## 4.1.2
 
 ### Patch Changes