+++ packages/callout/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leafygreen-ui/callout
 
+## 12.1.3
+
+### Patch Changes
+
+- 9cf3b18: Updates provider peer dependency version string to correctly use `pnpm` `workspace` syntax
+- Updated dependencies [9cf3b18]
+  - @leafygreen-ui/typography@22.2.2
+
 ## 12.1.2
 
 ### Patch Changes