+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @leafygreen-ui/modal
 
+## 20.3.2
+
+### Patch Changes
+
+- 587f14e: Add explicit `margin: auto;` to `<dialog>` element styles to prevent `margin` property being reset
+
+## 20.3.1
+
+### Patch Changes
+
+- 9cf3b18: Updates provider peer dependency version string to correctly use `pnpm` `workspace` syntax
+- Updated dependencies [9cf3b18]
+  - @leafygreen-ui/icon-button@17.1.3
+
 ## 20.3.0
 
 ### Minor Changes