+++ packages/stepper/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leafygreen-ui/stepper
 
+## 6.1.3
+
+### Patch Changes
+
+- 9cf3b18: Updates provider peer dependency version string to correctly use `pnpm` `workspace` syntax
+- Updated dependencies [9cf3b18]
+  - @leafygreen-ui/typography@22.2.2
+  - @leafygreen-ui/tooltip@14.2.3
+
 ## 6.1.2
 
 ### Patch Changes